[IMP] food: improved the structure of delivery_partner
Removed the delegation inheritance as it was creating multiple records for
delivery partner.

Added current order assigned to the delivery partner field
which will be ready only

Added and improved the compute methods in food_order on state and
delivery_partner_id change

Available delivery partner can be auto-assigned after accepting the order.